requirements.txt
Django==5.0.2
django-allauth==0.61.1
django-ipware==6.0.4
django-sendfile2==0.7.1
django-sslserver==0.22
gunicorn==21.2.0
phonenumbers==8.13.30
psycopg2-binary==2.9.9
reportlab==4.0.6
